Taught the project how to use (and possibly build) VecMem.
Tidied up the CMake configuration a bit, making sure that the settings
made in traccc could be in sync with the settings made in VecMem.

Introduced a couple of helper CMake modules for this, under the cmake/
subdirectory. This includes traccc-vecmem.cmake, which takes care of
either finding, or building VecMem (in case it was not found).

Updated the traccc::core library to "link against" vecmem::core.