Updated the algorithmic code to the VecMem based cell EDM.
The main re-design was done in the traccc::component_connection code.
That was re-written to have a public interface specifically for the
host and device versions of the cell EDM, while internally implementing
the connected component labelling using a single templated code.

All the other places in the code just needed the "obvious" updates
for instantiating cell objects using a memory resource, and to handle
the updated traccc::cell_container type correctly.