Implement container element views
The container classes were recently refactored into a more generalized
class, which allows us to write generalized element views for them.
Currently, we are accessing them directly through the inner vectors, but
this is dangerous and makes it extremely difficult to enforce invariants
that we would like to have for containers. This change introduces an
accessor view for header-vector elements in the containers, which allows
us to more safely and more ergonomically access the inner elements.
Effectively, the semantics are very similar to those of
array-of-structs, even though the underlying layout is a
struct-of-arrays.