Use HTTPS for data submodule instead of SSH
Currently, the submodule for the traccc data, which comes from CERN
Gitlab, is set to use SSH. However, this means that anyone who cannot
log in to CERN Gitlab cannot fully check out this repository. By
changing to the HTTPS protocol, we resolve this issue, since the data
repository is public. This should make it easier for external developers
to work on traccc, and it will make working with CI easier (see #50).