Add Python-based CCL data generator tool
Like I presented a few weeks ago in the parallelization meeting, I have
been using my own data generator to test CCL and CCA implementations. In
case it is useful to anyone else, I wanted to keep it around in the
tracc repository. The data generated by this tool is devoid of any and
all physical meaning, but it is specifically designed to expose the
performance-sensitive parameters of the input data. In this case, that
is the number of modules, the size of each modules, and the parameters
to the log-normal distribution governing the number of hits per module,
and the number of activations per hit.