Add ACTS as a dependency through FetchContent
As we have seen in #35, we need to somehow include ACTS as a dependency
for traccc so we can compile it easily, including on the CI. To make
this possible, this pull request adds some CMake code to fetch and build
the ACTS core in the same way that we build the vecmem code right now.
That is to say, we see if it is installed on the system. If it is, we're
done. If it isn't, CMake fetches the code from source for us, and adds
it into the tree for building. This should give us a flexible solution
to the dependency problem, at least for now.