Reject warnings in CI builds
As we have discussed in #64, we want to reject warnings in CI builds in
order to prevent potential code smells from creeping into our code
base. This commit enables the -Werror flag on all CI builds, and adds
some CMake logic to ensure that the dependencies can still be built
correctly.