Require CMake 3.18 or higher for builds with CUDA
As it turns out CMake versions 3.17 and lower are not able to build CUDA
correctly with the C++17 standard. To ensure that builds work correctly,
I am adding a check to ensure that the minimum version is 3.18, which
doe support C++17.

This commit should clear the CI after acts-project/machines#38 is
merged.