Updated the compiler flag settings in the same way as it's done in all other projects.
The C++, CUDA and SYCL settings have been separated into different files,
and these are only used in the subdirectories that need them.

Note that "-Werror all-warnings" was not added for CUDA yet, because a
warning is still showing up from detray, which will need a bit of thinking
to solve properly.