Remove dependency on ACTS unit test headers
Right now, our tests are relying on some ACTS-internal headers which are
defined for the ACTS unit tests. This does not work, because those
headers are not internal, and they are not exported or installed,
leading to compiler errors when building traccc. This commit is a
stop-gap solution that copies over the necessary files into the domain
of traccc such that the code can compile.