Fix range-loop-construct warning in module map
Recent versions of gcc introduce a new `range-loop-construct` warning
which detects non-explicit copies or references in for loops. This
started throwing warnings for a particular loop in our module map, which
then got escalated to an error. This commit fixes the problem and allows
the code to compile with gcc 11.2.0.