Migrate README flowchart to Mermaid
In the spirit of "it's Saturday", I decided to migrate the flowchart of
algorithms that we have in our README to Mermaid, which is a graph
markup language that recently gained support in Github. It looks pretty
and best of all, you can click on links to go places. It's great.