Implement a CUDA version of the module map
The module map was originally designed to be GPU friendly, and we now
reap the fruits of that decision. This implementation of a CUDA version
of that code is quite short and simple to understand. This commit adds
an owning CUDA version of the module map, as well as a light-weight
non-owning view object which is designed to be passed to kernels.