Add tests for the CUDA module map
This test is rather simple, but effective. It loads a CPU module map
from an input file, constructs a CUDA module map and then - for each
module - launches a kernel which retrieves the relevant transformation
matrix and sends it back to the CPU. The CPU will then check whether
these transformation matrices retrieved from the GPU are correct.