Introduced helper macros for declaring standard container types.
For each data type we need to declare a large number of C++ types
in the code. The new macros are meant to ensure that those container
types would be declared consistently throughout the project.

At the same time ensured that container views could be freely
constructed in device code. (Which apparently we didn't do in CUDA
code yet.)