Improve memory performance of CUDA CCA code
As it stands, the CUDA CCA code uses managed memory because it's the
path of least resistance. Sadly, it is also the path of least
performance. In this commit, we change the CUDA CCA code to use device
memory rather than managed memory, and this speeds up the algorithm
significantly. In my preliminary measurements, the difference is roughly
a factor five.

Also changes the code to use the new `traccc::memory_resource` type.