Container Comparison Update, main branch (2022.10.26.) (#259)
* Introduced a new way of comparing containers.

Added it in addition for comparing spacepoints in the SYCL
algorithm sequence. With the goal of adding it for a bunch
of additional comparisons as well later on.

* Created a new/updated traccc::performance library.

Merged the code from the existing traccc::performance interface
library and the newly introduced object/container comparison code
into a single place. All while slightly updating the code.
(Moved container_comparator into the traccc:: namespace
directly.)

* Introduced an additional layer of indirection in container_comparator.

This additional indirection will be necessary for comparing seeds
in an elegant way, it was really not necessary for the spacepoints
themselves.

At the same time started with a bit of cleanup in the SYCL algorithm
sequence example.

* Added a coparison for the reconstructed track parameters.

Introduced traccc::container_comparator, and defined a comparison
for traccc::bound_track_parameters.

* Introduced a comparison for seeds.

And with that in place, updated both the CUDA and SYCL seeding and
"full" sequence example applications to use the new code for
measuring the agreement between the host and device results.

* Made the normalisation of the agreement values more conservative.

Now they are normalised wrt. the larger container of the two being
compared.

* Re-implemented traccc::details::is_same_angle more like tha ATLAS implementation.

* Added a comparison of the surface links of track parameters.

And to make sure that they would be the same between the host
and device code, I had to make sure that the device code would
call on the constructor of the EDM class.