Memory Resource Cleanup (Part 1), main branch (2022.10.26.) (#261)
* Made traccc::sycl::clusterization_algorithm expect device memory as input.

Very sneakily the clusterization code so far was running off of
host memory. To fix it, and only use vecmem::host_memory_resource
in the SYCL example (instead of vecmem::sycl::host_memory_resource),
the memory management was re-shuffled a bit.

At the same time updated traccc_seedfinding_example_sycl as well
to use separate host and device memory instead of shared memory.

* Made traccc::cuda::clusterization_algorithm expect device memory as input.

Making the same update in it, and in the CUDA example applications,
as the ones made in the SYCL algorithm and applications as well.