Re-wrote the binary file writing code.
Moved the code itself into a compiled source file, and
turned traccc::io into a shared library.

Hardcoded the use of std::size_t for the size variables in
both the writing and the reading code. This seems a bit
safer than using the size types from the specific containers.