Implement block-wide odd-even sort
Sorting small arrays is a relatively common problem in GPGPU
programming. Many useful algorithms exist, and some are provided by
libraries like CUB. An algorithm close to my heart is odd-even sort
because it is exceedingly simply, relatively efficient for small arrays
and, importantly, it uses O(1) space. This commit adds new
implementations of this sorting algorithm for block-wide odd-even sort
in a portable way.