Implement near-transparent layout-swappable arrays
It is common in GPGPU programming to experiment with struct-of-arrays
(SoA) and array-of-structs (AoS) layouts. Sadly, C++ does not natively
allow us to easily switch between these layouts, and doing so requires a
lot of programming. This commit adds a new class that allows
near-transparent switching of SoA and AoS layouts, relying on the
compiler being able to cleverly eliminate unwanted operations.