Implement warp ballot sync with indexing in CUDA
It suffices to say that inter-thread communication is an important
feature in GPGPU computing. One particularly common operation is to
count the number of threads for which some predicate evaluates to true
(for example, that predicate can determine whether to write something to
global memory). This is commonly done using shared memory, but this is
not necessary. CUDA provides ballot synchronization primitives, but this
does not give you the index of the current thread in the set of threads
that evaluate a predicate to true. This commit adds a new utility
function which adds this functionality. Given a predicate, this function
synchronizes a warp and returns a unique, consecutive index of the
current thread in the set of threads for which the predicate is true.