Add quality reporting for seeds of any length
As far as I understand, the seed efficiency tools we have now don't
really work with seeds that are not of size 3, and it also appears that
they are not currently working for the CUDA code. This PR introduces
code that functions similarly, but allows matching and filtering to be
configured. It also outputs its data to files that can be analyzed
further.