Reduce the memory usage of array wrapper tests
The array wrapper tests use a rather large amount of memory in order to
establish the performance relevance of the changes, but this also breaks
the CI tests in #237. This commit reduces the allocation size for
several gigabytes to several megabytes, which the CI machine should be
able to satisfy in virtually all situations.