Adds a digitization configurator that can read a digitization meta configuration and generate a full digitization configuration for a detray geometry. This is mostly useful for geometries that are generated in detray, as this should hopefully become compatible with ACTS generated tracking geometries at some point, in which case the detray surfaces will provide a source link to the corresponding ACTS surface.
The configuration can only be done per volume and extra bit. Since the current ACTS version
in traccc does not support extra-bit matching in the GeometryHierarchyMap yet,
and because detray has no facilities to set extra bits during geometry IO yet,
this part of the configuration has not been tested so far.