Renamed the experimental clusterization algorithm source file.
As it turns out, nvcc has a race condition with compiling two
source files called clusterization_algorithm.cu at the same
time (in Debug mode).

At the same time moved the unnamed namespaces to the global scope
in the two source files, so that the symbols would truly not be
exposed from the object files.