Made the CPU algorithm sequence use Detray geometries.
This is needed for processing data produced for the ODD geometry.

To be able to read those files correctly, taught the CSV cell
reading code how to switch from Acts to Detry identifiers in the
cell module data on the fly. To make the simulation files usable
with the identifiers generated by Detray.