Made digitization info reading work with geometry_id remapping.
The test files consistently contain "Acts geometry identifiers".
While the identifier needs to be remapped to the "Detray identifier"
when looking for the surface transforms, we still need the
"Acts identifier" to look up the digitization parameters of
the modules.

This is a bit hacky way of achieving this, but for now it
should be good enough.