Removed track finding and fitting from the full algorithm sequence examples.
Unfortunately that code does not yet work correctly, so let's not
add it yet.

Made it possible in all seq_example executables to read either the
"old style" geometry description of the TML layout, or to use
detray::detector to read in a geometry (and accompanying data file)
using Detray.

Unfortunately the SYCL and CUDA examples do not work with the ODD
test data. Something with clusterization is going wrong.