Update to detray v0.61.0 (#528)
This brings the new stepsize scaling and toy detector. In the new toy detector default configuration, I set the thickness of the homogeneous material slabs back to 1.5mm, which unfortunately causes the CKF test to fail and the simulator to get stuck. The same thing happens with the old toy detector, if the thickness is increased and the mapped material (on the portals) set to vacuum, so I don't believe it is a problem of the new implementation. Instead, I now configured the new toy detector to much less material budget (0.15mm as it was set on the old toy detector). On the CKF test, I even needed a little less. I am guessing, this could be due to the increased detector acceptance (some holes in the endcaps are gone)?

On this PR some CUDA CI problems were observed that seemed unrelated to the changes