Re-enabled reading duplicate cells from CSV files.
The truth mapping code (currently) relies on possibly having
duplicate cells in the event record, for different particles
depositing energy in the same cell. The truth mapping code,
and some of the I/O unit tests, are now set up to disable the
cell de-duplication during CSV reading.

Reverted to the previous ordering scheme for traccc::cell-s,
and implemented a (different) custom ordering for the cell
collection. Since as it turns out, the project implicitly
relies on these two behaving a little differently.