Made throughput measurements "ODD capable".
Taught the host and CUDA algorithms how to perform track
finding and fitting when a Detray geometry is used.

Updated the common code of the throughput applications to
deal with reading in a Detray geometry when appropriate,
and to create the processing algorithms with their new
interfaces.

Updated the SYCL algorithm to mimic the host and CUDA
algorithms. Even though it will not perform any track
finding or fitting for the moment.