Updated the GitLab CI builds to the latest Docker images.
At the same time introduced SYCL builds/tests with an Intel and
NVIDIA backend as well. (The Intel backend running on the CPU
on the test machine for the moment.) This required installing the
NVIDIA oneAPI plugin during the test, as that was left out of
the existing Acts Docker image. :-(