Improve numerical stability of CCL variance
I noticed that, at some point, a factor of $\frac{1}{12}$ was added to
the variance of measurements and this slipped through because the
tolerance on the variance test was extremely large, i.e. no smaller than
0.1. This is an unacceptably high tolerance, and so I decided that the
variance computation was in need of an update. I decided to adopt two
strategies to do this. The first is the implementation of Welford's
online algorithm, which relies on the following recurrence relation:

$$\sigma^2_n = (1 - \frac{w_n}{W_n}) \sigma^2_{n-1} + \frac{w_n}{W_n} *
(x_n - \mu_n) (x_n - \mu_{n-1})$$

This is significantly less prone to catastrophic cancellation. Second, I
shifted the entire computation by the position of the first cell, which
brings the computation closer to zero where floating point computation
is more accurate. This depends on two equivalences:

$$\mu(x_1, \ldots, x_n) = \mu(x_1 - C, \ldots, x_n - C) + C$$

and

$$\sigma^2(x_1, \ldots, x_n) = \sigma^2(x_1 - C, \ldots, x_n - C)$$

Combined, these factors allow me to drop the tolerance in the tests from
a _minimum_ of 0.1 to a fixed value of 0.0001.