Add SYCL tests for CCA code
Currently, we implement CCA tests only for CPU code and for CUDA code,
but we have an easily reusable test harness for this, so we might as
well test our SYCL code, too. This commit adds a new set of tests to the
SYCL suite which tests the functionality and accuracy of our CCL code.