Download data files after build in CI
As it stands right now, we download all the data files before we ever
even build the traccc project. This is a bit wasteful, as a failed build
will render all that downloading useless. This commit moves the data
downloading step to after the build step, and also makes it conditional
on whether the tests will actually be run.