Require multiple threads for Alpaka CCL code
The CCL code currently does not support running with a single thread per
block; it has to be run with multiple threads per block by design. In
order to require this, I have added a new function that asserts this
about a device (note that in Alpaka 1.2.0 there is `IsMultiThreadAcc`,
but we cannot use this yet). I have also added a static assertion for
this.