Fixed the I/O test failures.
One failure was because of a difference in the "module link"
variable of cells. Since in the new scheme this link would
mean an index in the "global" DD object.

The other failure was however because of a genuine bug
in the spacepoint reading that I introduced while updating
the I/O code to use the new DD type.