Removed the variance_y variable.
As it turned out, there was no reason to keep this variable in
memory. The values coming from the digitization files were the same
as what we get with the formula implemented in the measurement
creation code.

At the same time made the code for detecting 1D modules a lot more
robust. Now the decision is made based on the number of "bins" in
a module's segmentation.