Fix integer comparison bugs in device CKF
This commit fixes a few bugs in the device CKF to do with the integral
representation of the previous step number. Firstly, the previous step
of step 0 (e.g. some sort of sentinel value) was being inconsistently
defined as either the maximum value of a signed int or an unsigned int,
and this was stored in an integer, leading to unexpected behaviour.
Furthermore, the track building kernel was not correctly checking for
the sentinel value, but rather looking for values smaller than 0.