Introduce traccc::silicon_detector_description, main branch (2024.06.21.) (#627)
* Introduced "named detector types" to the project.

The traccc algorithms will need to work on a concrete list of detector
types, they cannot be fully templated.

Started with adding I/O for the types that support it.

* Starting with the re-write of the detector data handling.

Declared an SoA container called traccc::detector_description that
would hold all of the detector description data needed by the
measurement and spacepoint creation algorithms. Fillable both from
the old-style CSV files used for the TrackML detector, and from
JSON Detray geometry configuration files.

* Starting the move to traccc::detector_description.

Updated the host-based measurement creation algorithm
to use the new type.

* Moved to reading cells using the new detector description.

* Saving a snapshot of the code.

In a quite broken state...

* Made traccc::io compile against the new detector description.

* Made traccc_seq_example work with the new detector description.

* Made the host throughput executables work with the new detector description.

* Added module dimensionality info to the detector description.

While fixing the measurement creation logic for 1D modules,
and fixing the code formatting of the recently modified files.

* Made traccc::cuda and traccc::sycl build with traccc::detector_description.

At the same time harmonized the CUDA and SYCL code implementation
a little.

* Updated the cpu tests to traccc::detector_description.

* Updated the CUDA tests to traccc::detector_description.

At the same time fixing some mistakes that I added while updating
the CUDA clusterization code for using traccc::detector_description.

* Made all host and CUDA example applications build.

Also ran some minimal tests with them, but nothing super
thorough.

* Made all the SYCL example and test code build.

With practically no runtime testing for now.

* Updated the I/O code for the latest Detray version.

* Made traccc_par_example work with the detector description.

At the same time updated the code to use the same command line
setup as all the other executables do.

Removed the other OpenMP example, as it was not actually doing
anything useful anyway...

* Made all the Alpaka example and test code build.

* Made the benchmarks work.

At the same time simplified the benchmarking code
a little.

* Made the Kokkos example build.

* Fixed the I/O test failures.

One failure was because of a difference in the "module link"
variable of cells. Since in the new scheme this link would
mean an index in the "global" DD object.

The other failure was however because of a genuine bug
in the spacepoint reading that I introduced while updating
the I/O code to use the new DD type.

* traccc::detector_description -> traccc::silicon_detector_description.

In preparation for introducing detector description objects for
other types of detectors as well later on.

* Simplify the creation of detray::tracking_surface.

Following Joana's recommendation.

* DD member naming changes.

Removed the variance_z variable, as it was not used anymore.
Named the "geometry identifier" variables as geometry_id()
and acts_geometry_id(), following Beomki's suggestion.

* Simplified the setting of the measurement dimensionality.

The code no longer uses branching to set the 2D coordinates
differently for 1D and 2D measurements. The second coordinate
is just assumed to be ignored for 1D measurements instead.