Attach algebra-plugins define to core target
Right now, the algebra plugin to be used is selected using a global
define in CMake, but this does not carry over when traccc is used as a
dependency, e.g. like we do in ACTS. This PR ensures that the define is
instead added as a target-specific public definition, which will resolve
this issue.