Switched to using edm::silicon_cell<T> proxy objects.
Also introduced the usage of detector description proxy
objects in a few places.

Updated the sanity functions once again, to bring them back
to a setup closer to how they are at the moment in the
main branch. Also synchornized the CUDA and SYCL implementations
of the sanity functions a little.