Named the host track finding traccc::host::ckf_algorithm.
Switched to implementing the reconstruction code for different
Detray detector and Covfie magnetic field types as overloads on
the call operator of a single "CKF algorithm".

Updated all clients to use this new name.

Split the source file, to parallelize compilation, and hopefully
reduce memory usage in individual build processes.