Fix register overload in CUDA spacepoint formation
PR #719 refactored the spacepoint formation to use the detray detector.
While this was and remains a good idea, the introduction of the detray
detector massively increases the register use of the kernel, to the
point that I was seeing "too many resources requested for launch"
errors. This commit modifies the kernel to specify appropriate launch
bounds, allowing the compiler to better tune the register usage and
resolve the bug.