Re-wrote spacepoint formation as a non-templated algorithm.
Renamed the algorithm to traccc::host::silicon_pixel_spacepoint_formation_algorithm,
to make it very clear what the algorithm is doing. Once we introduce logic
for creating spacepoints off of silicon strip measurements, we can see
what to do. (Rename this algorithm, or add a separate algorithm that does
the strip spacepoint creation.)