Remove Thrust pair type from prefix sum code
It turns out (see #816) that simply removing Thrust from the core
library is not enough; it needs to be further removed from the common
device code, too. Although this commit is unlikely to resolve the issue
(as some Thrust functions are still called in the common device code),
removing the last use of the Thrust pair type serves to make me feel
better.