Fix zero-track edge case bug in CUDA track finding
This commit fixes a bug discovered by Neza; if the track finding finds
exactly zero tips, it never sets the total number of track candidates to
zero, which causes it to use an uninitialized value in a later kernel.
This, in turn, leads to an assertion error. In this commit, we fix the
bug and add an assertion to protect against any future errors.

Co-authored-by: Neža Ribarič <neza.ribaric@cern.ch>