Add deterministic processing mode to examples
This commit adds an optional deterministic processing mode to the
throughput examples. It also makes the seed configurable. The default
behaviour of the examples remains as it is now, but passing the
`--deterministic` flag will simply cycle through events in order, and
the `--random-seed` flag allows users to configure their own random
seed. If the random seed is 0 (the default) then a time-based seed is
used instead.