Use configuration types as SSOT for defaults
As it stands, some of our default values for configuration parameters,
e.g. those of the track finding, have defaults defined both in the
configuration type as well as in the corresponding options type in the
examples. This is confusing as @paradajzblond discovered today. In this
commit, I remove the defaults defined in the examples code and treat the
configuration type defaults as the single source of truth for default
values.