Increase collection comparison performance
Although the collection comparison code is not performance-critical, it
is still very, very slow. In part, this is because it is written very
naively. This commit improves the performance of this code by several
orders of magnitude by sorting the collections, and by avoiding checks
for elements which have already been matched at a lower uncertainty
level. After these optimisations, it turned out that the Iterators-based
progress bar was the performance-limiting factor, so I had to remove
that as well.