ref: benchmark toy detector config (#714)
Allow the toy detector grids to find all neighboring surfaces, this should increase the number of tracks that can be found and fitted. Furthermore, I restricted the simulation to an eta region, where the tracks should hit a decent number of sensitive surface. Apart from that, the detector is now copied to device instead of using managed memory.

I also remove the copy of the track states back to host containers and allowed openMP to use dynamic scheduling for better load balancing