Allow event data reader to work on Geant4 data
Currently, the event data reader for performance-related examples (e.g.
the truth seeding) fails on Geant4 data as it is possible for Geant4 to
generate multiple measurements in the exact same place. This previously
caused an exception, but we can downgrade this to simply merge the
measurements. This will allow the truth examples to work on Geant4 data.