Add failure propagation to Kálmán filter code
This commit enables our Kálmán filter code to fail, e.g. it allows the
smoothers and updaters to return a success value, and that success value
is returned to the outside algorithm. This allows us to elegantly handle
edge cases like those described in #858, as well as cases where tracks
propagate exactly along the z-axis.

Co-authored-by: Beomki Yeo <beomki.yeo@gmail.com>