SoA Spacepoints, main branch (2025.02.19.) (#878)
* Introduced SoA spacepoints in traccc::core.

* Introduced SoA spacepoints in traccc::io.

* Simplify the API of traccc::device::form_spacepoints.

There is no need for it to receive the size of the measurement
collection separately. It already has a view of the collection
with that information.

* Make traccc::edm::spacepoint<T> use the point3 type natively.

Also, make at least some of the code use the new
vecmem::edm::host::push_back(...) functionality.

* Simplify(?) the for-loop.

* Don't use "auto" for spacepoints and seeds.

* Stop traccc::kokkos::details::spacepoint_binning from being an algorithm.

In sync with all the other spacepoint_binning classes.

* Update traccc::nseed_performance_writer.

Note though that the code is not actually used anywhere at the moment.
I only ensured that it would compile.