Improve assertions in Kálmán filter and fitter
This commit makes some improvements to the assertions in the Kálmán
filter and fitter. Namely, it asserts that the number of measurements
added in the track building kernel is the correct number, and it asserts
that none of the measurements are duplicated. In the fitting kernel, the
assertion on the fitting status is loosened.