Improve memory usage in track finding postamble
The code which turns the tips of our track finding into actual tracks
uses an excessive amount of memory, as it massively overallocates.
Indeed, it allocates memory as though all tips have the maximum number
of track states, which is unrealistic. This commit makes it so that the
number of valid track states is counted instead, making the allocation
more precise. In my measurements, this more than halves the memory usage
of traccc on $\langle\mu\rangle = 200$ ttbar events in the ODD.