Sort truth track measurements by time
Currently, our truth track measurements are unsorted, and this appears
to cause some problems with tracking, i.e. with seeds that start way too
far from the interaction point. This commit sorts measurements in our
truth data by the timing information in the measurement and, if that is
not available, the timing information in the hit data.