Fix bug in seed selection in track finder
PRs #911 and #915 inadvertently introduced a new bug in the track
finding where tracks on which the first measurement is a hole would
potentially cause an out-of-bounds access in the seed array. This commit
fixes this bug by always using the seed ID which is already encoded in
the candidate link.