Accelerate and fix bug in performance code
This commit accelerates some of the performance code and fixes a bug in
it. The bug fix concerns very low eta tracks which would be put into the
non-existant bin -1. The optimizations are simple (using references and
precomputing some mappings which cause the performance comparison code
to speed up by a factor 20 for $\mu = 80$ events.