Default to zeroth device, if non given.
Alpaka, unlike CUDA, doesn't have an equivalent of "Get Current Device",
such that we can't really use that.

This is hopefully a reasonable middle-ground, as if a user does care
about what device they are using, they should pass a device over.

Otherwise, they get a reasonable default (and per device environment
arguments can be used to change this, as they would for native CUDA, HIP
& SYCL code.