Fix dangling reference in full chain algorithms
A segmentation fault spotted by @CrossR has exposed a few problems in
the way we handle full chain algorithms in our examples code.
Fundamentally, the problem is that we create a full chain algorithm `A`
which contains a copy object `A.copy` and a subalgorithm `A.subalg`.
This subalgorithm has a copy object reference `A.subalg.copy = A.copy`.
We then move `A` into `A'` which properly moves `A.copy` into `A'.copy`
and `A.subalg` into `A'.subalg`, but the reference `A'.subalg.copy`
still points to `A.copy` and not `A'.copy`, leading to a dangling
reference.

I fixed this problem by using smart pointers, which avoid having the
copy object embedded in the algorithm. I also had to make some further
changes to make the algorithms properly movable, but this allowed me to
remove some of the very verbose custom move- and copy constructors we
had!