Alpaka: Vecmem Interface (#960)
* Introduce traccc::alpaka::queue and traccc::alpaka::vecmem_objects.

Meant as an example/guidance for Ryan Cross.

* Update examples and tests.

Needs testing on both HIP and SYCL still.

* Drop old code + guard around HIP copy for now.

Can be removed when vecmem is updated.

* Default to zeroth device, if non given.

Alpaka, unlike CUDA, doesn't have an equivalent of "Get Current Device",
such that we can't really use that.

This is hopefully a reasonable middle-ground, as if a user does care
about what device they are using, they should pass a device over.

Otherwise, they get a reasonable default (and per device environment
arguments can be used to change this, as they would for native CUDA, HIP
& SYCL code.

* Fix how throughput examples use copy.

* Swap back to HIP async copy, now available.

* Start fixing PR comments.

* Fix tests, add optional alpaka queue input for traccc::alpaka::queue.

* Update for PR comments.

* Special case SYCL for getting device ID.

* Drop storing the device in Alpaka Queue wrapper.

---------

Co-authored-by: Attila Krasznahorkay <Attila.Krasznahorkay@cern.ch>
Co-authored-by: Attila Krasznahorkay <30694331+krasznaa@users.noreply.github.com>