Compile fitter only in direction-matching kernel
This is the first optimization made possible by #994. It ensures that
the forward fitter actor is only compiled into the forward kernel, and
that the backward fitter is only compiled in the backward kernel. This
reduces the size of the kernels by about 30%, thus hopefully improving
compile times and instruction cache hit rates.