Select first rather than last smoothed parameter
The final step of the Kálmán fitter algorithm collects one of the
smoothed parameters. It is supposed to collect the first one, but the
current implementation collects the last one due to a missing break
statement. This commit adds the missing statement, ensuring that we
always pick the first smoothed state.