Reduce the number of smoothed parameter copies
The Kálmán fitter algorithm finds the first smoothed track state by
iterating over all of them and copying each of them. This invokes a lot
of unnecessary copies. To alleviate this, this commit ensures that only
one track parameter is copied.

Although this code is not particularly hot, this change might bring a
small performance improvement.