(CUDA) Templating Rewrite, main branch (2025.06.20.) (#1029)
* Re-wrote the CUDA CKF algorithm.

So that it would follow the same UI that all the other CKF algorithms
provide.

Had to modify the kernel specialization code a little, as it was itself
relying on definitions from the old algorithm class. In sort of a
circular way.

* Adapting the client code to the CUDA CKF rewrite.

* Re-wrote the CUDA KF algorithm.

So that it would follow the same UI that all the other KF algorithms
provide.

* Adapting the client code to the CUDA KF rewrite.

* Introduced a common set of typedefs for the CKF algorithms.

While also re-designing the templating in the functions that implement
the CKF algorithms for the different backends.

* Introduced a common typedef for the KF algorithms.

While also slightly re-designing the templating in the functions that
implement the KF algorithms for the different backends.