Implement new doublet cut based on triplet radii
This commit adds a new cut to the doublet finding which is able to
determine based on only the doublets if it is possible to find a third
spacepoint which will form a seed within the required helix radius
bounds.

This is the cut discussed on the traccc Mattermost recently.