Inhomogeneous Magnetic Field, main branch (2025.06.23.) (#1033)
* Switch to the v10 data file(s).

* Introduced a generic traccc::bfield type.

Along with an additional type definition for inhomogeneous
fields to use in host code.

* Introduced I/O code for inhomogeneous magnetic fields.

Along with a trivial unit test for the code.

* Added inhomogeneous field host algorithms.

* Adapted all clients to the modified host algorithm APIs.

* Added inhomogeneous field CUDA algorithms.

* Adapted all clients to the modified CUDA algorithm APIs.

* Added inhomogeneous field SYCL algorithms.

* Adapted all clients to the modified SYCL algorithm APIs.

* Added inhomogeneous field support to the Alpaka algorithms.

But didn't actually implement the support in the device code yet.

* Adapted all clients to the modified Alpaka algorithm APIs.

* Made the GitLab CI use Ubuntu 24.04 based images.

The same ones that the GitLab CI has been using since a while.
In order to avoid issues with too old GCC versions.