Add duplicate track removal to the CKF algorithm
this commit introduces a new stage to the existing CKF code which is
designed to filter out duplicate tracks. It works by finding tracks at
every step which have the same spacepoints and filtering them out. This
leads to a reduction in the total number of tracks propagated and
therefore to a reduction in computation time.