Add pre-commit check for duplicate CUDA files
For reasons of CUDA being CUDA, we currently cannot reliable compile the
project in debug mode if there are multiple CUDA files with the same
base name. This causes a race condition during compilation. In order to
relieve this issue, this commit adds a pre-commit check that verifies
that all CUDA file base names are unique.