Enable fast math library in CUDA compilation
One of the biggest performance bottlenecks of our code right now are the
math functions we use from the standard library. To illustrate the
problem, these functions have the problem that they contain a _lot_ of
edge case handling code for NaN values, infinities, etc. which are not
necessary for our use case. This causes an incredible amount of code to
be added and causes a huge amount of instruction cache misses when
though CUDA contains native intrinsics of high accurary which model the
same behaviour. This PR serves to illustrate the massive performance
gain that we can achieve by using intrinsic math rather than the math
provided by the C++ standard library.