Rework the maximum branch per seed cut
This commit reworks the CKF cut on the maximum number of branches per
initial seed. The current implementation makes little sense as it resets
its internal counter in every step, which means that we can actually
create far more tracks than the cut would suggest. This is therefore a
borderline bugfix.