Embed bottom and middle spacepoint in triplet
Right now, we store only the top spacepoint in the device triplet which
is technically all that is required, but this makes it very difficult
and time-consuming to retrieve the bottom and middle spacepoint. This
commit embeds the locations of those spacepoints in the triplet, making
life easier for a lot of the planned seeding changes.