Rework seed confirmation counting mechanism
Currently, we could seed confirmations and add them to the weight of the
seed. We add a large weight, 200, which causes several problems with the
seeding code. Most prominently, our weights before the confirmation are
in the range [-10, 0] which means that adding 200 brings you in the
range [190, 200] and the minimum weight is 200. Thus, you are very
coarsely cutting seeds on a very high minimum weight. Expressing the
confirmations using an integer counting system makes us more robust
against the aforementioned.