Track Fit/State SoA, main branch (2025.08.05.) (#1112)
* Introduced an SoA EDM for the track fit in traccc::core.

* Adapted traccc::device_common and traccc::cuda to the SoA track fit EDM.

* Adapted traccc::alpaka and traccc::sycl to the SoA track fit EDM.

* Adapted the tests and benchmarks to the SoA track fit EDM.

* Adapted the examples to the SoA track fit EDM.

* Introduced track_candidates_container::device and track_fit_container::device.

To make it a bit easier to deal with the slightly complex EDMs
of track finding and fitting in the code.

* Moved "complex EDM operations" to helper functions.

* Adjusted the host Kalman fit testing code a little.

Taking into account that the updated code now returns tracks
that have not been successfully fitted. Previously it would not
return the tracks for which the fit failed.

* Reintroduce print_fitted_tracks_statistics(...).

But now as a function in the examples code. Not as part of
traccc::core.

* Small changes in the host fitting code.

Based on the PR feedback.

* Made the plotting tool documentation more explicit.