Fix and rework Kálmán filter p and pT cuts
This commit fixes the $p$ and $p_T$ cuts in the Kálmán filter, which
were previously entirely broken. The CLI options could not correctly
distinguish between the $p$ and $p_T$ cut, a distinction which is
entirely unnecessary in the first place. This commit splits the cut up
so that both can be supported at the same time, and sets a sensible
default for the $p_T$ cut.