Avoid async size copies into pageable memory
Launching asynchronous copies into pageable memory is highly
inefficient in CUDA, as it requires the entire CUDA context to lock,
across multiple threads. Now that our kernels are running as fast as
they are, this is starting to have a noticable impact on our
performance. This commit avoids the copies into pageable memory by using
some pinned memory for memory copies of sizes.