Fix invalid state containers in CUDA examples
Unfortunately it seems that I was way too quick to merge in #1178 as I
am now discovering that that PR contains some serious design flaws. In
this particular case, the default constructor for the new "track
container" allows for such a container to be created without any
measurements, which makes it extremely easy to create a track container
in an invalid state, i.e. with states containing valid measurement IDs
but without any measurements at all. This commit fixes this issue in two
of the CUDA examples. It is likely that the issue exists in many other
examples, but these are the most pressing to fix.