update to detray v0.104.1 (#1166)
In the new detray version, the mask tolerances are scaled according to the positional and directional covariances found on the previous surfaces (scaled with the distance to the next surface, but not transported with a full Jacobian). This will allow to open (and close) the surface finding tolerances dynamically, according to how well known the track model is. Since this can lead to including more surfaces in the CKF on which ultimately no fitting measurement could be found (and hence would lead to the loss of the track due to an increased hole count), the navigator now knows if a surface was only included due to an "edge hit" in the tolerance band. In this case, no hole is flagged and the track can be kept alive. Any data like this, which needs to be retained after the propagation is halted, has been added to the new propagation_data struct that is passed between the different kernels. The last and second last sensitive visited is added to the CKF-aborter, so that no propagation loops on now overlapping surfaces occur.

Edit: The mask tolerance scaling is turned off for the CKF but enabled for the KF (by the defaults in the propagation config).