Regularize the track covariances in the Kalman filter and patch strip measurement handling (#1184)
Add the check for negative diagonal elements to the KF. If the variance is just slightly negative, take the absolute value, otherwise throw an error. The error is a new Kalman fitter status "ERROR_UPDATER_INVALID_COVARIANCE" or "ERROR_SMOOTHER_INVALID_COVARIANCE".

The requirements for the smoothed chi2 have been relaxed, so that the fit will only be aborted if it is below -10.

Edit: Also contains a hotfix for the strip measurement projector matrix, which will get properly fixed in a later PR